paletteに関連するファイルを切り出す
